fix type page
