+++ apps/web/app/future/auth/logout/page.tsx
@@ -1,6 +1,7 @@
-import Logout from "@pages/auth/logout";
 import { withAppDirSsr } from "app/WithAppDirSsr";
 import { WithLayout } from "app/layoutHOC";
+
+import Logout from "@components/pages/auth/logout";
 
 import { getServerSideProps } from "@server/lib/auth/logout/getServerSideProps";
