+++ apps/web/app/future/auth/signin/page.tsx
@@ -1,13 +1,14 @@
-import signin from "@pages/auth/signin";
 import { withAppDirSsr } from "app/WithAppDirSsr";
 import { WithLayout } from "app/layoutHOC";
-import type { InferGetServerSidePropsType } from "next";
+
+import signin from "@components/pages/auth/signin";
 
 import { getServerSideProps } from "@server/lib/auth/signin/getServerSideProps";
+
+const getData = withAppDirSsr(getServerSideProps);
 
 export default WithLayout({
   getLayout: null,
   Page: signin,
-  // @ts-expect-error TODO: fix this
-  getData: withAppDirSsr<InferGetServerSidePropsType<typeof getServerSideProps>>(getServerSideProps),
+  getData,
 })<"P">;
