+++ apps/web/components/pages/team/[slug]/[type].tsx
@@ -4,19 +4,13 @@ import { Booker } from "@calcom/atoms";
 import { getBookerWrapperClasses } from "@calcom/features/bookings/Booker/utils/getBookerWrapperClasses";
 import { BookerSeo } from "@calcom/features/bookings/components/BookerSeo";
 
+import type { AppProps } from "@lib/app-providers";
 import { type PageProps } from "@lib/team/[slug]/[type]/getServerSideProps";
 
-export default function Type({
-  slug,
-  user,
-  booking,
-  away,
-  isEmbed,
-  isBrandingHidden,
-  entity,
-  duration,
-  isInstantMeeting,
-}: PageProps) {
+const Type: React.FC<PageProps> & {
+  PageWrapper?: AppProps["Component"]["PageWrapper"];
+  isBookingPage?: boolean;
+} = ({ slug, user, booking, away, isEmbed, isBrandingHidden, entity, duration, isInstantMeeting }) => {
   return (
     <main className={getBookerWrapperClasses({ isEmbed: !!isEmbed })}>
       <BookerSeo
@@ -41,4 +35,6 @@ export default function Type({
       />
     </main>
   );
-}
+};
+
+export default Type;
