+++ apps/web/pages/team/[slug]/[type].tsx
@@ -1,14 +1,9 @@
-import { type AppProps } from "@lib/app-providers";
 import { getServerSideProps, type PageProps } from "@lib/team/[slug]/[type]/getServerSideProps";
 
-import PageWrapper, { type CalPageWrapper } from "@components/PageWrapper";
-import TypePage from "@components/pages/team/[slug]/[type]";
+import PageWrapper from "@components/PageWrapper";
+import Type from "@components/pages/team/[slug]/[type]";
 
 export { getServerSideProps, type PageProps };
-
-const Type = TypePage as unknown as CalPageWrapper & {
-  isBookingPage: AppProps["Component"]["isBookingPage"];
-};
 
 Type.PageWrapper = PageWrapper;
 Type.isBookingPage = true;
